Extract functions from `explicit-length-check.js`
