+++ rules/explicit-length-check.js
@@ -2,6 +2,7 @@
 const {isParenthesized} = require('eslint-utils');
 const getDocumentationUrl = require('./utils/get-documentation-url');
 const isLiteralValue = require('./utils/is-literal-value');
+const {isBooleanNode, getBooleanAncestor} = require('./boolean');
 
 const TYPE_NON_ZERO = 'non-zero';
 const TYPE_ZERO = 'zero';
@@ -12,23 +13,6 @@ const messages = {
 	[MESSAGE_ID_SUGGESTION]: 'Replace `.length` with `.length {{code}}`.'
 };
 
-const isLogicNot = node =>
-	node &&
-	node.type === 'UnaryExpression' &&
-	node.operator === '!';
-const isLogicNotArgument = node =>
-	isLogicNot(node.parent) &&
-	node.parent.argument === node;
-const isBooleanCall = node =>
-	node &&
-	node.type === 'CallExpression' &&
-	node.callee &&
-	node.callee.type === 'Identifier' &&
-	node.callee.name === 'Boolean' &&
-	node.arguments.length === 1;
-const isBooleanCallArgument = node =>
-	isBooleanCall(node.parent) &&
-	node.parent.arguments[0] === node;
 const isCompareRight = (node, operator, value) =>
 	node.type === 'BinaryExpression' &&
 	node.operator === operator &&
@@ -71,23 +55,6 @@ const lengthSelector = [
 	'[property.type="Identifier"]',
 	'[property.name="length"]'
 ].join('');
-
-function getBooleanAncestor(node) {
-	let isNegative = false;
-	// eslint-disable-next-line no-constant-condition
-	while (true) {
-		if (isLogicNotArgument(node)) {
-			isNegative = !isNegative;
-			node = node.parent;
-		} else if (isBooleanCallArgument(node)) {
-			node = node.parent;
-		} else {
-			break;
-		}
-	}
-
-	return {node, isNegative};
-}
 
 function getLengthCheckNode(node) {
 	node = node.parent;
@@ -133,37 +100,6 @@ function getLengthCheckNode(node) {
 	}
 
 	return {};
-}
-
-function isBooleanNode(node) {
-	if (
-		isLogicNot(node) ||
-		isLogicNotArgument(node) ||
-		isBooleanCall(node) ||
-		isBooleanCallArgument(node)
-	) {
-		return true;
-	}
-
-	const {parent} = node;
-	if (
-		(
-			parent.type === 'IfStatement' ||
-			parent.type === 'ConditionalExpression' ||
-			parent.type === 'WhileStatement' ||
-			parent.type === 'DoWhileStatement' ||
-			parent.type === 'ForStatement'
-		) &&
-		parent.test === node
-	) {
-		return true;
-	}
-
-	if (parent.type === 'LogicalExpression') {
-		return isBooleanNode(parent);
-	}
-
-	return false;
 }
 
 function create(context) {
